+++ src/locales/en/index.js
@@ -2,12 +2,8 @@
 // Licensed under the MIT License
 
 export default {
-  indexTitle: 'Front Page',
-  indexServicesId: 'services',
-  indexServicesTitle: 'What Do We Refract?',
-  indexReadMore: 'Read more',
-  indexPortfolioId: 'portfolio',
-  indexPortfolioTitle: 'Portfolio',
-  indexContactId: 'contact',
-  indexContactTitle: 'Contact',
+  indexContactFormName: 'Name',
+  indexContactFormEmail: 'Email',
+  indexContactFormTel: 'Phone Number',
+  indexContactFormMessage: 'Your message for us',
 };